requirements.txt
jupyter
notebook
ipykernel
ipython
nbconvert
nbformat
jupyterlab
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
scikit-image
opencv-python
dlib
imutils
tqdm
ipywidgets
Flask==2.1.2
Pillow==10.1.0
python-dotenv==1.0.1
line-bot-sdk
gunicorn==20.1.0